Added logic to save the access token after logging in
After logging in using the api/auth/token the accesstoken from the response will be stored in the authStore.
The authStore has a ref accesstoken to the localstorage. changed the httpClient also so that the authorization header also includes this token.
When the user is not logged in the default value of the key will be null